[Bugfix] Add bounds check for SHMManager handle in get_singleton_instance

+++ csrc/cpu/shm.cpp
@@ -292,7 +292,7 @@ class SHMManager {
 
   static SHMManager* get_singleton_instance(int64_t handle) {
     TORCH_CHECK(handle >= 0
-                    && handle < static_cast<int64_t>(SingletonInstances.size()),
+                    && static_cast<size_t>(handle) < SingletonInstances.size(),
                 "Invalid SHMManager handle: ", handle);
     return SingletonInstances[handle].get();
   }
